Verify button (hidden)
Add a hidden verify button, that will read out the result and compare
filament consumption with one of the original file, thus verifying there
is no systematic retract/unretract inconsistency.

The need for it arose from Repetier-Host displaying different filament
consumption for original and processed. It seems to be a repetier-host
problem.